feat: add depthLimit to route handler
