--- src/routes/graphql/types/subscription.ts
@@ -1,4 +1,0 @@
-export interface ISubscription {
-  userId: string;
-  authorId: string;
-}
